componetized out CardsMap to display cards by type
